pyproject.toml
[project]
name = "nornir-nokia-srlinux-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncssh==2.22.0",
    "bcrypt==5.0.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "deepdiff>=8.6.1",
    "dictdiffer==0.9.0",
    "dotenv==0.9.9",
    "fastapi>=0.128.0",
    "gnmi==0.0.0a0",
    "grpcio==1.76.0",
    "idna==3.11",
    "iniconfig==2.3.0",
    "invoke==2.2.1",
    "jinja2==3.1.6",
    "junos-eznc==2.7.5",
    "logging>=0.4.9.6",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mcp>=1.25.0",
    "mdurl==0.1.2",
    "napalm==5.1.0",
    "napalm-srl==1.0.5",
    "ncclient==0.7.0",
    "netaddr==1.3.0",
    "netmiko==4.6.0",
    "netutils==1.15.2",
    "nornir==3.5.0",
    "nornir-jinja2==0.2.0",
    "nornir-napalm==0.5.0",
    "nornir-netbox==0.3.0",
    "nornir-netmiko==1.0.1",
    "nornir-pygnmi==0.2.0",
    "nornir-rich==0.2.0",
    "nornir-scrapli==2025.1.30",
    "nornir-srl>=0.3.5",
    "nornir-utils==0.2.0",
    "ntc-templates==8.1.0",
    "packaging==25.0",
    "paramiko==4.0.0",
    "pip==25.3",
    "pluggy==1.6.0",
    "protobuf==3.20.3",
    "pycparser==2.23",
    "pyeapi==1.0.4",
    "pygments==2.19.2",
    "pygnmi==0.8.15",
    "pynacl==1.6.2",
    "pynetbox==7.5.0",
    "pyparsing==3.3.1",
    "pyserial==3.5",
    "pytest==9.0.2",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rich==13.9.4",
    "ruamel-yaml==0.19.1",
    "scp==0.15.0",
    "scrapli==2025.1.30",
    "scrapli-cfg==2025.1.30",
    "scrapli-community==2025.1.30",
    "scrapli-netconf==2025.1.30",
    "setuptools==80.9.0",
    "six==1.17.0",
    "textfsm==2.1.0",
    "transitions==0.9.3",
    "ttp==0.10.0",
    "ttp-templates==0.3.7",
    "typing-extensions==4.15.0",
    "urllib3==2.6.2",
    "yamlordereddictloader==0.4.2",
]
